add "noSync" variants of changeNodePart
These variants can be used in cases where neither the gain cache nor
attributed gains are needed. This is currently only the case in initial
partitioning, but it could also be useful for synchronous refinement
algorithms (such as Jet).

Unfortunately, the current implementation requires that
`resetEdgeSynchronization` is called afterwards to maintain some
invariants.