use mini_boost target (#177)
When you use the KAHYPAR_DOWNLOAD_BOOST cmake option without an existing boost installation (what seems to be the intended use case to me), cmake fails afterwards because Boost::program_options is not found.

With these changes, this is fixed. As a side effect, the download option now precedes both existing installations and the disable option introduced in #170.