introduce a more uniform handling of contraint violations (#225)
The goal is to implement all logic for evaluating constraint violations
(balance, empty blocks and others) in a single place (metrics.h), along
with the comparison logic to decide which of two partitions is
preferable. Then, all IP and refinement algorithms can simply use this
instead of always duplicating the logic -- which gets quite tricky if
more constraints than just balance are involved.